jonas är bäst

+++ src/main/java/com/ailab/tools/LaserPropertiesResponse.java
@@ -16,10 +16,10 @@ public class LaserPropertiesResponse implements Response
       Map<String, Object> pose = (Map<String, Object>)data.get("Pose");
       Map<String, Object> orientation = (Map<String, Object>)pose.get("Orientation");
       
-      double w =  (Double)orientation.get("W");      
-      double x =  (Double)orientation.get("X");
-      double y =  (Double)orientation.get("Y");
-      double z =  (Double)orientation.get("Z");
+      int w =  (Integer)orientation.get("W");
+      int x =  (Integer)orientation.get("X");
+      int y =  (Integer)orientation.get("Y");
+      int z =  (Integer)orientation.get("Z");
       
       return new double[] {w, x, y, z};
    }
